refactor(provider-bindgen): fix import & export semantics
The current semantics are unintuitive -- interfaces
"exported" from a provider signifies functionality *provided* to the
lattice, and the WIT worlds should reflect that.

At present, functionality exposed to the lattice for actors to use is
signified by WIT *imports* in the provider worlds.

This commit reverses the import and export semantics that were used
for WIT contracts in the providers, to make them more intuitive.
