+++ crates/providers/blobstore-s3/wit/provider-blobstore.wit
@@ -1,6 +1,6 @@
 package wasmcloud:provider-blobstore-s3;
 
 world provider {
-    import wasmcloud:blobstore/blobstore;
-    export wasmcloud:blobstore/chunk-receiver;
+    import wasmcloud:blobstore/chunk-receiver;
+    export wasmcloud:blobstore/blobstore;
 }
