+++ crates/providers/lattice-controller/wit/provider-lattice-controller.wit
@@ -1,5 +1,5 @@
 package wasmcloud:provider-lattice-controller;
 
 world provider {
-    import wasmcloud:lattice-control/lattice-controller;
+    export wasmcloud:lattice-control/lattice-controller;
 }
