+++ crates/providers/messaging-nats/wit/provider.wit
@@ -1,6 +1,6 @@
 package wasmcloud:provider-messaging-nats;
 
 world provider-messaging-nats {
-    import wasmcloud:messaging/messaging;
-    export wasmcloud:messaging/message-subscriber;
+    import wasmcloud:messaging/message-subscriber;
+    export wasmcloud:messaging/messaging;
 }
